Update dungeonGenerationAlgorithms.py
Changed libtcod to tcod since I have it straight up installed. Changed the xrange instances to range since python 3 and newer just use range instead. Added a bunch of int() to math functions that were returning floats. Changed the CellularAutomata class function connectCaves function because in <if (newDistance < distance) or distance == None:> the distance variable was set to None and python couldn't compare none to an integer. Changed the distance variable to 0 and the == to 0 as well. Other than those things this is sooooo sweet, it helped me so much!